requirements.txt
torchvision~=0.15.2
tensorboardx~=2.6.1
pillow~=9.4.0
numpy~=1.24.3
tqdm~=4.65.0